Initial support for SQL v130 client tools stack. This supports both SQL Server 2016 and the latest Azure SQL DB features.
Changes to refresh support for current features require updates to the strongly typed model - this is to be expected as the intention is to block if required methods are missing. In this case the BodyDependencies relationship added to ColumnStoreIndex was updated as this had been added in recent releases.

Note on DacFx bugs: It was discovered during this update that features such as Row Level Security were not shown in the public model. This will be fixed in an upcoming release and processes were added to ensure this does not occur in the future.